working on validation and controllers

+++ app/Http/Controllers/SeniorController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Senior;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class SeniorController extends Controller
 {
@@ -15,10 +17,20 @@ class SeniorController extends Controller
      */
     public function index()
     {
-        return response()
-            ->json(User::with('senior', 'adresse_habitation', 'senior.forfait')
-            ->has('senior')
-            ->get());
+        if(Auth::check())
+        {
+            $user = Auth::user();
+
+            if ($user->can('voir-liste-seniors'))
+            {
+                return response()
+                    ->json(User::with('senior', 'adresse_habitation', 'senior.forfait')
+                        ->has('senior')
+                        ->get(), Response::HTTP_OK);
+            }
+        }
+
+        return response()->json(['error' => 'Unauthorized'],Response::HTTP_UNAUTHORIZED);
     }
 
     /**
@@ -29,7 +41,17 @@ class SeniorController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        if(Auth::check())
+        {
+            $user = Auth::user();
+
+            if ($user->can('creer-senior'))
+            {
+                return response()->json("", Response::HTTP_OK);
+            }
+        }
+
+        return response()->json(['error' => 'Unauthorized'],Response::HTTP_UNAUTHORIZED);
     }
 
     /**
@@ -40,10 +62,20 @@ class SeniorController extends Controller
      */
     public static function show($id)
     {
-        return response()
-            ->json(User::with('senior', 'adresse_habitation', 'senior.forfait')
-                ->has('senior')
-                ->find($id));
+        if(Auth::check())
+        {
+            $user = Auth::user();
+
+            if ($user->can('voir-senior'))
+            {
+                return response()
+                    ->json(User::with('senior', 'adresse_habitation', 'senior.forfait')
+                        ->has('senior')
+                        ->find($id));
+            }
+        }
+
+        return response()->json(['error' => 'Unauthorized'],Response::HTTP_UNAUTHORIZED);
     }
 
 
@@ -56,7 +88,17 @@ class SeniorController extends Controller
      */
     public function update(Request $request, Senior $senior)
     {
-        //
+        if(Auth::check())
+        {
+            $user = Auth::user();
+
+            if ($user->can('modifier-senior'))
+            {
+                return response()->json("", Response::HTTP_OK);
+            }
+        }
+
+        return response()->json(['error' => 'Unauthorized'],Response::HTTP_UNAUTHORIZED);
     }
 
     /**
@@ -67,10 +109,28 @@ class SeniorController extends Controller
      */
     public function destroy($id)
     {
-        $senior = Senior::find($id);
-        if(isset($senior)) {
-            $senior->delete();
+        if(Auth::check())
+        {
+            $user = Auth::user();
+
+            if ($user->can('supprimer-senior'))
+            {
+                $senior = Senior::find($id);
+
+                if(isset($senior))
+                {
+                    $senior->delete();
+
+                    return response()->json(['error' => 'Ok'], Response::HTTP_OK);
+                }
+                else
+                {
+                    return response()->json(['error' => 'Bad Request'], Response::HTTP_BAD_REQUEST);
+                }
+            }
         }
+
+        return response()->json(['error' => 'Unauthorized'],Response::HTTP_UNAUTHORIZED);
     }
 
     public static function get($id)
